ODS.sample_... like ODS.physics_... and ODS.plot_...
