+++ omas/omas_utils.py
@@ -6,105 +6,6 @@ import sys
 # --------------------------------------------
 # ODS utilities
 # --------------------------------------------
-def ods_sample():
-    """
-    create sample ODS data
-    """
-    from .omas_core import ODS
-    ods = ODS()
-
-    #check effect of disabling dynamic path creation
-    try:
-        ods.dynamic_path_creation = False
-        ods['info.user']
-    except LookupError:
-        ods['info'] = ODS()
-        ods['info.user'] = unicode(os.environ['USER'])
-    else:
-        raise(Exception('OMAS error handling dynamic_path_creation=False'))
-    finally:
-        ods.dynamic_path_creation = True
-
-    #check that accessing leaf that has not been set raises a ValueError, even with dynamic path creation turned on
-    try:
-        ods['info.machine']
-    except ValueError:
-        pass
-    else:
-        raise(Exception('OMAS error querying leaf that has not been set'))
-
-    # info ODS is used for keeping track of IMAS metadata
-    ods['info.machine'] = 'ITER'
-    ods['info.imas_version'] = default_imas_version
-    ods['info.shot'] = 1
-    ods['info.run'] = 0
-
-    # check .get() method
-    assert (ods.get('info.shot') == ods['info.shot'])
-    assert (ods.get('info.bad', None) is None)
-
-    # check that keys is an iterable (so that Python 2/3 work the same way)
-    keys = ods.keys()
-    keys[0]
-
-    # check that dynamic path creation during __getitem__ does not leave empty fields behind
-    try:
-        print(ods['wall.description_2d.0.limiter.unit.0.outline.r'])
-    except ValueError:
-        assert 'wall.description_2d.0.limiter.unit.0.outline' not in ods
-
-    ods['equilibrium']['time_slice'][0]['time'] = 1000.
-    ods['equilibrium']['time_slice'][0]['global_quantities']['ip'] = 1.5
-
-    ods2 = copy.deepcopy(ods)
-    ods2['equilibrium']['time_slice'][1] = ods['equilibrium']['time_slice'][0]
-    ods2['equilibrium.time_slice.1.time'] = 2000.
-
-    ods2['equilibrium']['time_slice'][2] = copy.deepcopy(ods['equilibrium']['time_slice'][0])
-    ods2['equilibrium.time_slice[2].time'] = 3000.
-
-    printd(ods2['equilibrium']['time_slice'][0]['global_quantities'].location, topic='sample')
-    printd(ods2['equilibrium']['time_slice'][2]['global_quantities'].location, topic='sample')
-
-    ods2['equilibrium.time_slice.1.global_quantities.ip'] = 2.
-
-    # uncertain scalar
-    ods2['equilibrium.time_slice[2].global_quantities.ip'] = ufloat(3,0.1)
-
-    # uncertain array
-    ods2['equilibrium.time_slice[2].profiles_1d.q'] = uarray([0.,1.,2.,3.],[0,.1,.2,.3])
-
-    # check different ways of addressing data
-    printd(ods2['equilibrium.time_slice']['1.global_quantities.ip'], topic='sample')
-    printd(ods2[['equilibrium', 'time_slice', 1, 'global_quantities', 'ip']], topic='sample')
-    printd(ods2[('equilibrium', 'time_slice', '1', 'global_quantities', 'ip')], topic='sample')
-    printd(ods2['equilibrium.time_slice.1.global_quantities.ip'], topic='sample')
-    printd(ods2['equilibrium.time_slice[1].global_quantities.ip'], topic='sample')
-
-    ods2['equilibrium.time_slice.0.profiles_1d.psi'] = numpy.linspace(0, 1, 10)
-
-    # pprint(ods.paths())
-    # pprint(ods2.paths())
-
-    # check data slicing
-    printd(ods2['equilibrium.time_slice[:].global_quantities.ip'], topic='sample')
-
-    ckbkp = ods.consistency_check
-    tmp = pickle.dumps(ods2)
-    ods2 = pickle.loads(tmp)
-    if ods2.consistency_check != ckbkp:
-        raise (Exception('consistency_check attribute changed'))
-
-    # check flattening
-    tmp = ods2.flat()
-    # pprint(tmp)
-
-    # check deepcopy
-    ods3=ods2.copy()
-
-    return ods3
-
-
 def different_ods(ods1, ods2):
     """
     Checks if two ODSs have any difference and returns the string with the cause of the different
@@ -225,7 +126,7 @@ def json_dumper(obj):
         return OrderedDict(zip(obj.keys(), obj.values()))
 
     tmp=is_uncertain(obj)
-    if any(numpy.atleast_1d(tmp)):
+    if numpy.any(numpy.atleast_1d(tmp)):
         if not len(numpy.array(tmp).shape):
             return dict(__ufloat__=nominal_values(obj),
                         __ufloat_std__=std_devs(obj))
